move mock_btc_client.go
